+++ ui/client/components/graph/EspNode/misc.ts
@@ -1,0 +1,12 @@
+export const rectWidth = 300;
+export const rectHeight = 60;
+export const nodeLabelFontSize = 15;
+export const edgeStroke = '#B3B3B3';
+export const maxLineLength = 24;
+export const maxLineCount = 2;
+export const summaryCountConfig = {
+    breakPoint: 5,
+    maxExtraDigits: 2,
+    fontSizeStep: 3.5,
+    defaultFontSize: 20,
+};
