+++ ui/client/index.d.ts
@@ -23,6 +23,11 @@ declare module "*.less" {
   export default classes
 }
 
+declare module "*.html" {
+  const content: string
+  export default content
+}
+
 declare module "*.svg" {
   const content: string
   export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>
